+++ packages/project-integrity/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/project-integrity
 
+## 1.0.16
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.11 → 2.1.12
+
+## 1.0.15
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.10 → 2.1.11
+
 ## 1.0.14
 
 ### Patch Changes